迁移Counter 组件到containers中
